+++ packages/db-indexing/.env.sample
@@ -1,1 +1,2 @@
 OPENAI_API_KEY=''
+NODE_ENV=development
