--- packages/db-indexing/src/environment.ts
@@ -1,3 +1,0 @@
-const isDevelopment = process.env.NODE_ENV !== 'production';
-
-export const __DEV__ = isDevelopment;
