+++ packages/db-indexing/src/global.d.ts
@@ -1,0 +1,1 @@
+declare var __DEV__: boolean;
